added username to comments by including user model in the included co…

+++ public/js/blogPost.js
@@ -31,16 +31,21 @@ const createComment = async () => {
 
 const deletePost = async () => {
   try {
-    const post_id = await getId();
-    console.log("This is post_id", post_id);
-    const response = await fetch(`/api/user/blog/${post_id}`, {
-      method: "DELETE",
-    });
-    if (response.ok) {
-      document.location.replace("/dashboard");
-      alert("Post Deleted!");
+    const confirmDelete = confirm("Are you sure you want to delete this post?");
+    if (confirmDelete) {
+      const post_id = await getId();
+      console.log("This is post_id", post_id);
+      const response = await fetch(`/api/user/blog/${post_id}`, {
+        method: "DELETE",
+      });
+      if (response.ok) {
+        document.location.replace("/dashboard");
+        alert("Post Deleted!");
+      } else {
+        alert("Failed to delete post.");
+      }
     } else {
-      alert("Failed to delete post.");
+      return;
     }
   } catch (err) {
     console.log(err);
